Try to fix CI issues
