+++ pyodide-build/pyodide_build/common.py
@@ -4,8 +4,10 @@ import os
 import re
 import subprocess
 import sys
+import textwrap
 from collections.abc import Generator, Iterable, Iterator, Mapping
 from pathlib import Path
+from typing import NoReturn
 
 import tomli
 from packaging.tags import Tag, compatible_tags, cpython_tags
@@ -316,3 +318,13 @@ def replace_env(build_env: Mapping[str, str]) -> Generator[None, None, None]:
     finally:
         os.environ.clear()
         os.environ.update(old_environ)
+
+
+def exit_with_stdio(result: subprocess.CompletedProcess[str]) -> NoReturn:
+    if result.stdout:
+        print("  stdout:")
+        print(textwrap.indent(result.stdout, "    "))
+    if result.stderr:
+        print("  stderr:")
+        print(textwrap.indent(result.stderr, "    "))
+    raise SystemExit(result.returncode)
