+++ pyodide-build/pyodide_build/out_of_tree/venv.py
@@ -4,7 +4,7 @@ import sys
 from pathlib import Path
 from textwrap import dedent
 
-from ..common import get_make_flag, get_pyodide_root, in_xbuild_env
+from ..common import exit_with_stdio, get_make_flag, get_pyodide_root, in_xbuild_env
 
 
 def main(parser_args: argparse.Namespace) -> None:
@@ -83,6 +83,9 @@ def run(dest: Path) -> None:
         capture_output=True,
         encoding="utf8",
     )
+    if result.returncode != 0:
+        print("ERROR: failed to invoke Pyodide")
+        exit_with_stdio(result)
 
     pymajor = get_make_flag("PYMAJOR")
     pyminor = get_make_flag("PYMINOR")
@@ -119,10 +122,6 @@ def run(dest: Path) -> None:
         + "\n"
     )
     pip_path.chmod(0o777)
-    print("pip:")
-    print("=" * 10)
-    print(pip_path.read_text())
-    print("=" * 10)
 
     other_pips = [
         bin / "pip3",
@@ -135,5 +134,11 @@ def run(dest: Path) -> None:
         pip.symlink_to(pip_path)
 
     toload = ["micropip"]
-    subprocess.run([bin / "pip", "install", *toload])
-    print("Successfully created pyodide virtual environment!")
+    subprocess.run(
+        [bin / "pip", "install", *toload],
+    )
+    if result.returncode != 0:
+        print("ERROR: failed to invoke pip")
+        exit_with_stdio(result)
+
+    print("Successfully created Pyodide virtual environment!")
